requirements.txt
PySimpleGUI==4.40.0
PyMuPDF==1.18.15
Pillow==8.2.0
